Changed template version to vs 2017
