+++ graphics/lab-1/lab-1.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31005.135
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.1382
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "lab-1", "lab-1.vcxproj", "{BC6C5C7A-9381-4121-8950-4E587DC96593}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "lab-1", "lab-1\lab-1.vcxproj", "{B3590D2F-A650-48A4-8D13-97CCACED6AB4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,35 +17,35 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Debug|ARM.ActiveCfg = Debug|ARM
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Debug|ARM.Build.0 = Debug|ARM
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Debug|ARM.Deploy.0 = Debug|ARM
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Debug|ARM64.Build.0 = Debug|ARM64
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Debug|ARM64.Deploy.0 = Debug|ARM64
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Debug|x64.ActiveCfg = Debug|x64
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Debug|x64.Build.0 = Debug|x64
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Debug|x64.Deploy.0 = Debug|x64
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Debug|x86.ActiveCfg = Debug|Win32
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Debug|x86.Build.0 = Debug|Win32
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Debug|x86.Deploy.0 = Debug|Win32
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Release|ARM.ActiveCfg = Release|ARM
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Release|ARM.Build.0 = Release|ARM
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Release|ARM.Deploy.0 = Release|ARM
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Release|ARM64.ActiveCfg = Release|ARM64
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Release|ARM64.Build.0 = Release|ARM64
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Release|ARM64.Deploy.0 = Release|ARM64
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Release|x64.ActiveCfg = Release|x64
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Release|x64.Build.0 = Release|x64
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Release|x64.Deploy.0 = Release|x64
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Release|x86.ActiveCfg = Release|Win32
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Release|x86.Build.0 = Release|Win32
-		{BC6C5C7A-9381-4121-8950-4E587DC96593}.Release|x86.Deploy.0 = Release|Win32
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Debug|ARM.ActiveCfg = Debug|ARM
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Debug|ARM.Build.0 = Debug|ARM
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Debug|ARM.Deploy.0 = Debug|ARM
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Debug|ARM64.Build.0 = Debug|ARM64
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Debug|ARM64.Deploy.0 = Debug|ARM64
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Debug|x64.ActiveCfg = Debug|x64
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Debug|x64.Build.0 = Debug|x64
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Debug|x64.Deploy.0 = Debug|x64
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Debug|x86.ActiveCfg = Debug|Win32
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Debug|x86.Build.0 = Debug|Win32
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Debug|x86.Deploy.0 = Debug|Win32
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Release|ARM.ActiveCfg = Release|ARM
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Release|ARM.Build.0 = Release|ARM
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Release|ARM.Deploy.0 = Release|ARM
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Release|ARM64.ActiveCfg = Release|ARM64
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Release|ARM64.Build.0 = Release|ARM64
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Release|ARM64.Deploy.0 = Release|ARM64
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Release|x64.ActiveCfg = Release|x64
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Release|x64.Build.0 = Release|x64
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Release|x64.Deploy.0 = Release|x64
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Release|x86.ActiveCfg = Release|Win32
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Release|x86.Build.0 = Release|Win32
+		{B3590D2F-A650-48A4-8D13-97CCACED6AB4}.Release|x86.Deploy.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7CC9FE2C-7B0F-4A5B-81E9-0A1D2DE30E75}
+		SolutionGuid = {BCA3B6B4-3457-4B88-80D1-F4919D1FF986}
 	EndGlobalSection
 EndGlobal
