+++ graphics/lab-1/lab-1/Content/SamplePixelShader.hlsl
@@ -1,11 +1,11 @@
-// Per-pixel color data passed through the pixel shader.
+// Цветовые данные отдельных вершин, пропущенные через шейдер пикселей.
 struct PixelShaderInput
 {
 	float4 pos : SV_POSITION;
 	float3 color : COLOR0;
 };
 
-// A pass-through function for the (interpolated) color data.
+// Сквозная функция для (интерполированных) данных о цвете.
 float4 main(PixelShaderInput input) : SV_TARGET
 {
 	return float4(input.color, 1.0f);
