+++ graphics/lab-1/lab-1/Content/SampleVertexShader.hlsl
@@ -1,4 +1,4 @@
-// A constant buffer that stores the three basic column-major matrices for composing geometry.
+// Буфер констант, в котором хранятся три базовых матрицы в порядке по столбцам для составления геометрии.
 cbuffer ModelViewProjectionConstantBuffer : register(b0)
 {
 	matrix model;
@@ -6,33 +6,33 @@ cbuffer ModelViewProjectionConstantBuffer : register(b0)
 	matrix projection;
 };
 
-// Per-vertex data used as input to the vertex shader.
+// Данные для отдельных вершин, используемые для передачи в шейдер вершин.
 struct VertexShaderInput
 {
 	float3 pos : POSITION;
 	float3 color : COLOR0;
 };
 
-// Per-pixel color data passed through the pixel shader.
+// Цветовые данные отдельных вершин, пропущенные через шейдер пикселей.
 struct PixelShaderInput
 {
 	float4 pos : SV_POSITION;
 	float3 color : COLOR0;
 };
 
-// Simple shader to do vertex processing on the GPU.
+// Простой шейдер для обработки вершин на графическом процессоре.
 PixelShaderInput main(VertexShaderInput input)
 {
 	PixelShaderInput output;
 	float4 pos = float4(input.pos, 1.0f);
 
-	// Transform the vertex position into projected space.
+	// Преобразование положения вершины в пространстве проекции.
 	pos = mul(pos, model);
 	pos = mul(pos, view);
 	pos = mul(pos, projection);
 	output.pos = pos;
 
-	// Pass the color through without modification.
+	// Сквозная передача цвета без изменения.
 	output.color = input.color;
 
 	return output;
