+++ graphics/lab-1/lab-1/lab-1.vcxproj.filters
@@ -1,66 +1,66 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <Filter Include="Common">
-      <UniqueIdentifier>bc6c5c7a-9381-4121-8950-4e587dc96593</UniqueIdentifier>
+    <Filter Include="Общие">
+      <UniqueIdentifier>b3590d2f-a650-48a4-8d13-97ccaced6ab4</UniqueIdentifier>
     </Filter>
-    <Filter Include="Assets">
-      <UniqueIdentifier>d7aff874-2ef8-4105-91df-df86a23fc504</UniqueIdentifier>
+    <Filter Include="Ресурсы">
+      <UniqueIdentifier>97b0553f-9da2-48b2-a220-6ba3bd6dc0b0</UniqueIdentifier>
       <Extensions>bmp;fbx;gif;jpg;jpeg;tga;tiff;tif;png</Extensions>
     </Filter>
-	<Filter Include="Content">
-      <UniqueIdentifier>5e5e7a2b-69b8-4d31-ab54-de6c7e24fec6</UniqueIdentifier>
+	<Filter Include="Содержимое">
+      <UniqueIdentifier>2ff1a22a-4a03-44e6-b9e6-09f7ab783129</UniqueIdentifier>
     </Filter>
     <ClInclude Include="Common\DirectXHelper.h">
-      <Filter>Common</Filter>
+      <Filter>Общие</Filter>
     </ClInclude>
 	<ClInclude Include="Common\StepTimer.h">
-      <Filter>Common</Filter>
+      <Filter>Общие</Filter>
     </ClInclude>
     <ClInclude Include="Common\DeviceResources.h">
-      <Filter>Common</Filter>
+      <Filter>Общие</Filter>
     </ClInclude>
     <ClCompile Include="Common\DeviceResources.cpp">
-      <Filter>Common</Filter>
+      <Filter>Общие</Filter>
     </ClCompile>
 	<ClInclude Include="Content\Sample3DSceneRenderer.h">
-      <Filter>Content</Filter>
+      <Filter>Содержимое</Filter>
     </ClInclude>
     <ClInclude Include="Content\SampleFpsTextRenderer.h">
-      <Filter>Content</Filter>
+      <Filter>Содержимое</Filter>
     </ClInclude>
     <ClInclude Include="Content\ShaderStructures.h">
-      <Filter>Content</Filter>
+      <Filter>Содержимое</Filter>
     </ClInclude>
 	<ClCompile Include="Content\Sample3DSceneRenderer.cpp">
-      <Filter>Content</Filter>
+      <Filter>Содержимое</Filter>
     </ClCompile>
     <ClCompile Include="Content\SampleFpsTextRenderer.cpp">
-      <Filter>Content</Filter>
+      <Filter>Содержимое</Filter>
     </ClCompile>
 	<FxCompile Include="Content\SamplePixelShader.hlsl">
-      <Filter>Content</Filter>
+      <Filter>Содержимое</Filter>
     </FxCompile>
     <FxCompile Include="Content\SampleVertexShader.hlsl">
-      <Filter>Content</Filter>
+      <Filter>Содержимое</Filter>
     </FxCompile>
     <Image Include="Assets\LockScreenLogo.scale-200.png">
-      <Filter>Assets</Filter>
+      <Filter>Ресурсы</Filter>
     </Image>
     <Image Include="Assets\SplashScreen.scale-200.png">
-      <Filter>Assets</Filter>
+      <Filter>Ресурсы</Filter>
     </Image>
     <Image Include="Assets\Square150x150Logo.scale-200.png">
-      <Filter>Assets</Filter>
+      <Filter>Ресурсы</Filter>
     </Image>
     <Image Include="Assets\Square44x44Logo.scale-200.png">
-      <Filter>Assets</Filter>
+      <Filter>Ресурсы</Filter>
     </Image>
     <Image Include="Assets\Square44x44Logo.targetsize-24_altform-unplated.png">
-      <Filter>Assets</Filter>
+      <Filter>Ресурсы</Filter>
     </Image>
     <Image Include="Assets\Wide310x150Logo.scale-200.png">
-      <Filter>Assets</Filter>
+      <Filter>Ресурсы</Filter>
     </Image>
   </ItemGroup>
 </Project>
